🏷️(react) move SelectProps to index.tsx
It makes more sense for this type to be next to the Select component.
