+++ packages/react/src/components/Forms/Select/mono.tsx
@@ -1,41 +1,15 @@
-import React, {
-  forwardRef,
-  PropsWithChildren,
-  useEffect,
-  useState,
-} from "react";
+import React, { forwardRef, useEffect, useState } from "react";
 import { UseSelectStateChange } from "downshift";
-import { FieldProps } from ":/components/Forms/Field";
 import { optionToValue, SubProps } from ":/components/Forms/Select/mono-common";
 import { SelectMonoSearchable } from ":/components/Forms/Select/mono-searchable";
 import { SelectMonoSimple } from ":/components/Forms/Select/mono-simple";
-import { SelectHandle } from ":/components/Forms/Select/index";
+import { SelectHandle, SelectProps } from ":/components/Forms/Select";
 
 export interface Option {
   value?: string;
   label: string;
   disabled?: boolean;
 }
-
-export type SelectProps = PropsWithChildren &
-  FieldProps & {
-    label: string;
-    hideLabel?: boolean;
-    options: Option[];
-    searchable?: boolean;
-    name?: string;
-    defaultValue?: string | number | string[];
-    value?: string | number | string[];
-    onChange?: (event: {
-      target: { value: string | number | undefined | string[] };
-    }) => void;
-    onBlur?: (event: {
-      target: { value: string | number | undefined | string[] };
-    }) => void;
-    disabled?: boolean;
-    clearable?: boolean;
-    multi?: boolean;
-  };
 
 export const SelectMono = forwardRef<SelectHandle, SelectProps>(
   (props, ref) => {
